Fix Krazy2 qMin/qMax issues.

+++ src/libkst/vector.cpp
@@ -589,11 +589,11 @@ void Vector::oldChange(QByteArray &data) {
       qds >> _v[i];
       if(!i) {
           _min=_max=_minPos=sum=_v[i];
-          _minPos=qMax(_minPos,0.0);
+          _minPos=qMax(_minPos,qreal(0.0));
       } else {
           _min=qMin(_v[i],_min);
           _max=qMax(_v[i],_max);
-          _minPos=qMin(qMax(_v[i],0.0),_minPos);
+          _minPos=qMin(qMax(_v[i],qreal(0.0)),_minPos);
           sum+=_v[i];
       }
     }
@@ -620,11 +620,11 @@ void Vector::change(QByteArray &data) {
       qds >> _v[i];
       if(!i) {
           _min=_max=_minPos=sum=_v[i];
-          _minPos=qMax(_minPos,0.0);
+          _minPos=qMax(_minPos,qreal(0.0));
       } else {
           _min=qMin(_v[i],_min);
           _max=qMax(_v[i],_max);
-          _minPos=qMin(qMax(_v[i],0.0),_minPos);
+          _minPos=qMin(qMax(_v[i],qreal(0.0)),_minPos);
           sum+=_v[i];
       }
     }
